MAINT: move vectorization decorator to utility file
