+++ odl/space/fspace.py
@@ -27,7 +27,6 @@ from builtins import super
 import numpy as np
 from functools import wraps
 from inspect import isclass, isfunction
-from itertools import product
 
 # ODL imports
 from odl.operator.operator import Operator, _dispatch_call_args
@@ -35,8 +34,8 @@ from odl.set.sets import RealNumbers, ComplexNumbers, Set, Field
 from odl.set.space import LinearSpace, LinearSpaceVector
 from odl.util.vectorization import (
     is_valid_input_array, is_valid_input_meshgrid,
-    meshgrid_input_order, vecs_from_meshgrid,
-    out_shape_from_array, out_shape_from_meshgrid)
+    meshgrid_input_order, out_shape_from_array, out_shape_from_meshgrid,
+    vectorize)
 
 
 __all__ = ('FunctionSet', 'FunctionSetVector',
@@ -67,168 +66,6 @@ def _default_out_of_place(func, dtype, x, **kwargs):
     out = np.empty(out_shape, dtype=dtype)
     func(x, out=out, **kwargs)
     return out
-
-
-def vectorize(dtype=None, outarg='none'):
-    """Vectorization decorator for our input parameter pattern.
-
-    The wrapped function must be callable with one positional
-    parameter. Keyword arguments are passed through, hence positional
-    arguments with defaults can either be left out or passed by keyword,
-    but not by position.
-
-    Parameters
-    ----------
-    dtype : `type` or `str`, optional
-        Data type of the output array. Needs to be understood by the
-        `numpy.dtype` function. If not provided, a "lazy" vectorization
-        is performed, meaning that the results are collected in a
-        list instead of an array.
-    outarg : {'none', 'positional', 'optional'}
-        Type of the output argument of the decorated function for
-        in-place evaluation
-
-        'none': No output parameter. This is the default.
-        Resulting argspec: ``func(x, **kwargs)``
-        Returns: the new array
-
-        'positional': Required argument ``out`` at second position.
-        Resulting argspec: ``func(x, out=None, **kwargs)``
-        Returns: ``out``
-
-        'optional': optional argument ``out`` with default `None`.
-        Resulting argspec: ``func(x, out=None, **kwargs)``
-        Returns: ``out`` if it is not `None` otherwise a new array
-
-    Note
-    ----
-    For ``outarg`` not equal to 'none', the decorated function returns
-    the array given as ``out`` argument if it is not `None`.
-
-    Examples
-    --------
-    Vectorize a step function in the first variable:
-
-    >>> @vectorize(dtype=float)
-    ... def step(x):
-    ...     return 0 if x[0] <= 0 else 1
-
-    This corresponds to (but is much slower than)
-
-    >>> import numpy as np
-    >>> def step_vec(x):
-    ...     x0, x1 = x
-    ...     # np.broadcast is your friend to determine the output shape
-    ...     out = np.zeros(np.broadcast(x0, x1).shape, dtype=x0.dtype)
-    ...     idcs = np.where(x0 > 0)
-    ...     # Need to throw away the indices from the empty dimensions
-    ...     idcs = idcs[0] if len(idcs) > 1 else idcs
-    ...     out[idcs] = 1
-    ...     return out
-
-    Both versions work for arrays and meshgrids:
-
-    >>> x = np.linspace(-5, 13, 10, dtype=float).reshape((2, 5))
-    >>> x  # array representing 5 points in 2d
-    array([[ -5.,  -3.,  -1.,   1.,   3.],
-           [  5.,   7.,   9.,  11.,  13.]])
-    >>> np.array_equal(step(x), step_vec(x))
-    True
-
-    >>> x = y = np.linspace(-1, 2, 5)
-    >>> mg_sparse = np.meshgrid(x, y, indexing='ij', sparse=True)
-    >>> np.array_equal(step(mg_sparse), step_vec(mg_sparse))
-    True
-    >>> mg_dense = np.meshgrid(x, y, indexing='ij', sparse=False)
-    >>> np.array_equal(step(mg_dense), step_vec(mg_dense))
-    True
-
-    With output parameter:
-
-    >>> @vectorize(dtype=float, outarg='positional')
-    ... def step(x):
-    ...     return 0 if x[0] <= 0 else 1
-    >>> x = np.linspace(-5, 13, 10, dtype=float).reshape((2, 5))
-    >>> out = np.empty(5, dtype=float)
-    >>> step(x, out)  # returns out
-    array([ 0.,  0.,  0.,  1.,  1.])
-    """
-    def vect_decorator(func):
-
-        def _vect_wrapper_array(x, out, **kwargs):
-            # Assume that x is an ndarray
-            if out is None:
-                out_shape = out_shape_from_array(x)
-                if dtype is None:
-                    out = [0] * out_shape[0]
-                else:
-                    out = np.empty(out_shape, dtype=dtype)
-
-            for i, pt in enumerate(x.T):
-                out[i] = func(pt, **kwargs)
-            return out
-
-        def _vect_wrapper_meshgrid(x, out, **kwargs):
-            if out is None:
-                out_shape = out_shape_from_meshgrid(x)
-                if dtype is None:
-                    out = [0] * out_shape[0]
-                else:
-                    out = np.empty(out_shape, dtype=dtype)
-
-            order = meshgrid_input_order(x)
-            vecs = vecs_from_meshgrid(x, order=order)
-            for i, pt in enumerate(product(*vecs)):
-                out.flat[i] = func(pt, **kwargs)
-            return out
-
-        def _vect_wrapper(x, out, **kwargs):
-            # Find out dimension first
-            if isinstance(x, np.ndarray):  # array
-                if x.ndim == 1:
-                    dim = 1
-                elif x.ndim == 2:
-                    dim = len(x)
-                else:
-                    raise ValueError('only 1- or 2-dimensional arrays '
-                                     'supported.')
-            else:  # meshgrid
-                dim = len(x)
-
-            if is_valid_input_array(x, dim):
-                return _vect_wrapper_array(x, out, **kwargs)
-            elif is_valid_input_meshgrid(x, dim):
-                return _vect_wrapper_meshgrid(x, out, **kwargs)
-            else:
-                raise TypeError('invalid vectorized input type.')
-
-        @wraps(func)
-        def vect_wrapper_no_out(x, **kwargs):
-            if 'out' in kwargs:
-                raise TypeError("{}() got an unexpected keyword 'out'."
-                                "".format(func.__name__))
-            return _vect_wrapper(x, None, **kwargs)
-
-        @wraps(func)
-        def vect_wrapper_pos_out(x, out, **kwargs):
-            return _vect_wrapper(x, out, **kwargs)
-
-        @wraps(func)
-        def vect_wrapper_opt_out(x, out=None, **kwargs):
-            return _vect_wrapper(x, out, **kwargs)
-
-        outarg_ = str(outarg).lower()
-
-        if outarg_ == 'none':
-            return vect_wrapper_no_out
-        elif outarg_ == 'positional':
-            return vect_wrapper_pos_out
-        elif outarg_ == 'optional':
-            return vect_wrapper_opt_out
-        else:
-            raise ValueError('output arg type {!r} not understood.'
-                             ''.format(outarg))
-    return vect_decorator
 
 
 class FunctionSet(Set):
